requirements.txt
asyncpg
sqlalchemy[asyncio]
python-dotenv
asyncpg
jwt
pydantic-settings
pydantic[email]
httpx
langchain
langchain-core